arreglos

+++ Ferreteria Back-End/src/main/java/com/ferreteriaUniversal/ferreteria/repositorio/ItemProductoRepositorio.java
@@ -5,9 +5,9 @@ import java.util.List;
 import org.springframework.data.repository.Repository;
 import com.ferreteriaUniversal.ferreteria.model.ItemProducto;;
 
-public interface ItemProductoRepositorio extends Repository<ItemProducto, Integer> {
+public interface ItemProductoRepositorio extends Repository<ItemProducto, Long> {
 	List<ItemProducto> findAll();
-	ItemProducto findOne(int id);
+	ItemProducto findOne(long id);
 	ItemProducto save(ItemProducto itemProducto);
 	void delete(ItemProducto itemProducto);
